+++ addon/components/flexberry-edit-layer-feature.js
@@ -70,6 +70,16 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
   geoproviderMode: true,
 
   /**
+    Indicator for adding a layer to the map for edit mode.
+    When a layer is not activated in the layer tree
+
+    @property isLayerCopy
+    @type boolean
+    @default false
+  */
+  isLayerCopy: false,
+
+  /**
     Edit|Create|Union|Split|Diff|Import
   */
   mode: null,
@@ -139,7 +149,7 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
 
     this.set('error', null);
 
-    // Уберем редактирование с объектов, если оно было
+    // Cancel object editing, if there was one
     const layers = this.get('layers');
     if (layers) {
       Object.values(layers).filter((layer) => !isNone(layer)).forEach((layer) => {
@@ -167,7 +177,7 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
       this.set('initialData', {});
       this.set('mode', dataItems.mode);
 
-      this.set('choiceValueData', dataItems.choiceValueData); // эта штука должна быть только одна
+      this.set('choiceValueData', dataItems.choiceValueData); // there must be only one
 
       const editTools = this._getEditTools();
       set(leafletMap, 'editTools', editTools);
@@ -181,7 +191,7 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
           }
 
           if (this.get('state') === 'Edit') {
-            // сохраним геометрию, чтобы можно было быстро к ней вернуться при отмене
+            // save the geometry, to quickly return to it if the cancel action
             let latlngs;
             switch (layer.feature.geometry.type) {
               case 'Point':
@@ -211,7 +221,10 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
 
           const isMarker = layer instanceof L.Marker || layer instanceof L.CircleMarker;
 
+          // When a layer is not activated in the layer tree, it must be added to the leafletMap
+          // then turn on the layer editing mode
           if (!leafletMap.hasLayer(layer)) {
+            this.set('isLayerCopy', true);
             leafletMap.addLayer(layer);
           }
 
@@ -248,7 +261,7 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
 
         const { data, } = dataItem;
         this.set(`data.${index}`, data);
-        this.set(`initialData.${index}`, Object.assign({}, data)); // копия объекта для быстрого восстановления
+        this.set(`initialData.${index}`, Object.assign({}, data)); // A copy of the object for quick recovery
 
         index += 1;
       });
@@ -633,11 +646,11 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
     const initialDatas = this.get('initialData');
 
     if (!isNone(layers) && mode !== 'Saved') {
-      Object.keys(layers).forEach((index) => {
+      Ember.keys(layers).forEach((index) => {
         const layer = layers[index];
 
         if (mode === 'Edit') {
-          // отменим изменения в слое
+          // undo changes to the layer
           const latlng = latlngs[index];
 
           if (!isNone(layer) && !isNone(latlng)) {
@@ -660,7 +673,7 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
             }
           }
 
-          if (afterSave) { // если уходим после неудачного сохранения, то надо данные вернуть
+          if (afterSave) { // If we leave after a failed save, we have to return the data
             const data = initialDatas[index];
             if (!isNone(data)) {
               data.forEach((key) => {
@@ -673,33 +686,39 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
               });
             }
 
-            // для надписей
+            // for inscriptions
             leafletObject.editLayer(layer);
+          }
+
+          //Removing a layer from the map that was added for edit mode
+          if (this.get('isLayerCopy') && leafletMap.hasLayer(layer)) {
+            leafletMap.removeLayer(layer);
           }
 
           delete latlngs[index];
           delete layers[index];
           delete datas[index];
           delete initialDatas[index];
-        } else
-        // удалить слой
-        if (!isNone(layer)) {
-          if (!isNone(leafletObject) && leafletObject.hasLayer(layer)) {
-            leafletObject.removeLayer(layer);
-          }
-
-          if (leafletMap.hasLayer(layer)) {
-            leafletMap.removeLayer(layer);
-          }
-
-          const label = get(layer, '_label');
-          if (!isNone(label)) {
-            if (!isNone(leafletObject) && leafletObject.hasLayer(label)) {
-              leafletObject.removeLayer(label);
+        } else {
+          // remove the layer
+          if (!isNone(layer)) {
+            if (!isNone(leafletObject) && leafletObject.hasLayer(layer)) {
+              leafletObject.removeLayer(layer);
             }
 
-            if (leafletMap.hasLayer(label)) {
-              leafletMap.removeLayer(label);
+            if (leafletMap.hasLayer(layer)) {
+              leafletMap.removeLayer(layer);
+            }
+
+            let label = Ember.get(layer, '_label');
+            if (!isNone(label)) {
+              if (!isNone(leafletObject) && leafletObject.hasLayer(label)) {
+                leafletObject.removeLayer(label);
+              }
+
+              if (leafletMap.hasLayer(label)) {
+                leafletMap.removeLayer(label);
+              }
             }
           }
         }
@@ -707,12 +726,13 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
     }
 
     if (!isNone(layers)) {
-      // Сервисный слой общий с панелью атрибутов. Не надо очищать, если ничего не редактировали
+      // The service layer is shared with the attributes panel. No need to clear if you haven't edited anything
       this.send('clearSelected');
     }
 
     this.set('latlngs', null);
     this.set('layers', null);
+    this.set('isLayerCopy', false);
 
     this.set('data', null);
     this.set('initialData', null);
@@ -904,7 +924,7 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
 
           layer.disableEdit();
 
-          // создадим новый объект в слое
+          // Create a new object in the layer
           if (leafletObject.createLayerObject) {
             this.get('leafletMap').removeLayer(layer);
             layer = leafletObject.createLayerObject(leafletObject, data, layer.toGeoJSON().geometry);
@@ -965,6 +985,12 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
 
           layer.disableEdit();
           leafletObject.editLayer(layer);
+
+          // Deleting a copy of an edited layer from the map
+          if (this.get('isLayerCopy') && this.get('leafletMap').hasLayer(layer)) {
+            this.get('leafletMap').removeLayer(layer);
+            this.set('isLayerCopy', false);
+          }
         });
 
         event = 'flexberry-map:edit-feature';
@@ -990,7 +1016,7 @@ export default Component.extend(SnapDrawMixin, LeafletZoomToFeatureMixin, EditFe
         this.restoreLayers().then(() => {
           this.get('leafletMap').fire(`${event}:fail`, e);
         }).catch(() => {
-          // не удалось ни сохранить, ни восстановить слои. непонятно что делать
+          // can't save or restore layers. No decision on what to do next
           console.log('Save and restore layer error');
         });
       };