+++ addon/components/layers/wfs-layer.js
@@ -14,6 +14,7 @@ import jsts from 'npm:jsts';
 import BaseVectorLayer from '../base-vector-layer';
 import { checkMapZoom } from '../../utils/check-zoom';
 import { intersectionArea } from '../../utils/feature-with-area-intersect';
+import isUUID  from 'ember-flexberry-data/utils/is-uuid';
 import state from '../../utils/state';
 
 /**
@@ -233,7 +234,7 @@ export default BaseVectorLayer.extend({
             that.fire('error', {
               error: new Error(exceptionReport.message),
             });
-
+            reject(exceptionReport);
             return that;
           }
 
@@ -519,14 +520,34 @@ export default BaseVectorLayer.extend({
       const fieldsType = get(leafletObject, 'readFormat.featureType.fieldTypes');
       if (!isBlank(fieldsType)) {
         searchFields.forEach((field) => {
-          const typeField = fieldsType[field];
+          let typeField = fieldsType[field];
           if (!isBlank(typeField)) {
-            if (typeField !== 'string') {
-              equals.push(new L.Filter.EQ(field, e.searchOptions.queryString));
+            let accessProperty = false;
+            if (field !== 'primarykey') {
+              switch (typeField) {
+                case 'number':
+                  accessProperty = !e.context && !isNaN(Number(e.searchOptions.queryString));
+                  break;
+                case 'date':
+                  accessProperty = !e.context && new Date(e.searchOptions.queryString).toString() !== 'Invalid Date';
+                  break;
+                case 'boolean':
+                  accessProperty = !e.context && Boolean(e.searchOptions.queryString);
+                  break;
+                default:
+                  equals.push(new L.Filter.Like(field, '*' + e.searchOptions.queryString + '*', {
+                    matchCase: false
+                  }));
+                  break;
+              }
+
+              if (accessProperty && typeField !== 'string') {
+                equals.push(new L.Filter.EQ(field, e.searchOptions.queryString));
+              }
             } else {
-              equals.push(new L.Filter.Like(field, `*${e.searchOptions.queryString}*`, {
-                matchCase: false,
-              }));
+              if (isUUID(e.searchOptions.queryString)) {
+                equals.push(new L.Filter.EQ(field, e.searchOptions.queryString));
+              }
             }
           }
         });