+++ addon/components/map-commands-dialogs/export.js
@@ -1285,6 +1285,9 @@ const FlexberryExportMapCommandDialogComponent = Component.extend({
     onBeforeShow(e) {
       this.sendAction('beforeShow', e);
 
+      // Initialize default print/export options.
+      this.setDefaultExportOptions();
+
       // Switch scale control
       const leafletMap = this.get('leafletMap');
       const switchScaleControlMapName = this.get('switchScaleControlMapName');
@@ -2047,11 +2050,7 @@ const FlexberryExportMapCommandDialogComponent = Component.extend({
     this.set('_availablePaperFormats', A(Object.keys(paperFormatsInMillimeters)));
 
     // Initialize print/export options.
-    const defaultMapCaption = this.get('defaultMapCaption');
-    this.set('_options', $.extend(true, {}, defaultOptions, {
-      caption: defaultMapCaption,
-      fileName: defaultMapCaption,
-    }));
+    this.setDefaultExportOptions();
 
     // Bind context to tabular menu tabs 'click' event handler.
     this.set('actions.onSettingsTabClick', this.get('actions.onSettingsTabClick').bind(this));
@@ -2130,6 +2129,20 @@ const FlexberryExportMapCommandDialogComponent = Component.extend({
       this.set('_$tabularMenuTabItems', null);
       this.set('_$tabularMenuTabs', null);
     }
+  },
+
+  /**
+    When the form is expanded, the default settings are set
+
+    @method setDefaultExportOptions
+  */
+  setDefaultExportOptions() {
+    // Initialize default print/export options.
+    let defaultMapCaption = this.get('defaultMapCaption');
+    this.set('_options', $.extend(true, {}, defaultOptions, {
+      caption: defaultMapCaption,
+      fileName: defaultMapCaption,
+    }));
   },
 
   /**