+++ addon/layers/combine.js
@@ -30,7 +30,7 @@ export default BaseLayer.extend({
     Permitted operations related to layer type.
     @property operations
     @type String[]
-    @default ['edit', 'remove', 'identify', 'search', 'legend', 'attributes', 'filter']
+    @default ['edit', 'remove', 'identify', 'search', 'legend', 'attributes', 'editFeatures', 'filter']
   */
   operations: null,
 
@@ -39,7 +39,7 @@ export default BaseLayer.extend({
   */
   init() {
     this._super(...arguments);
-    this.operations = this.operations || ['edit', 'remove', 'identify', 'search', 'legend', 'attributes', 'filter'];
+    this.operations = this.operations || ['edit', 'remove', 'identify', 'search', 'legend', 'attributes', 'editFeatures', 'filter'];
   },
 
   /**